Single component for all Motion types.
